Re organize folder structure for chrome extension.
